+++ frontend/store/reducers/categories.ts
@@ -3,7 +3,7 @@ import {
     GET_CATEGORIES_SUCCESS,
     GET_CATEGORIES_ERROR
 } from "constants/categories";
-import { CategoriesAction, CategoriesState } from "types/categories";
+import { CategoriesAction, CategoriesState } from "types/state";
 
 export const initialState: CategoriesState = {
     data: [],
