+++ server/fixtures/catalogue.ts
@@ -1,4 +1,6 @@
-const catalogue = [
+import { ProductType } from "types/products";
+
+const catalogue: Array<ProductType> = [
     {
         "id": "1",
         "name": "branded shoes",
