+++ server/fixtures/categories.ts
@@ -1,4 +1,6 @@
-const categories = [
+import { CategoryType } from "types/categories";
+
+const categories: Array<CategoryType> = [
     { "name": "tshirts", "src": "img/categories/t-shirt.png" },
     { "name": "shoes", "src": "img/categories/shoes.png" },
     { "name": "accesories", "src": "img/categories/purse.png" }
