--- types/auth.d.ts
@@ -1,8 +1,0 @@
-import { ActionType } from "typesafe-actions";
-import { postAuthActions } from "actions/auth";
-import { FetchState } from ".";
-
-export type AuthAction = ActionType<typeof postAuthActions>;
-export type AuthState = {
-    readonly status: boolean;
-} & FetchState;
