+++ types/categories.d.ts
@@ -1,13 +1,6 @@
-import { ActionType } from "typesafe-actions";
-import { getCategoriesActions } from "actions/categories";
-import { FetchState } from ".";
-
-export type CategoryType = {
+type CategoryType = {
     name: string;
     src: string;
 };
 
-export type CategoriesAction = ActionType<typeof getCategoriesActions>;
-export type CategoriesState = {
-    readonly data: Array<CategoryType>;
-} & FetchState;
+export default CategoryType;
