+++ types/products.d.ts
@@ -1,8 +1,4 @@
-import { ActionType } from "typesafe-actions";
-import { getProductsActions } from "actions/products";
-import { FetchState } from ".";
-
-export type ProductType = {
+type ProductType = {
     category: string;
     id: string;
     name: string;
@@ -10,7 +6,4 @@ export type ProductType = {
     src?: string;
 };
 
-export type ProductsAction = ActionType<typeof getProductsActions>;
-export type ProductsState = {
-    readonly data: Array<ProductType>;
-} & FetchState;
+export default ProductType;
